+++ ors-engine/src/main/java/com/graphhopper/routing/ev/GoodsAccess.java
@@ -1,0 +1,13 @@
+package com.graphhopper.routing.ev;
+
+public class GoodsAccess {
+    public static final String KEY = "goods_access";
+
+    private GoodsAccess() {
+        // do not instantiate
+    }
+
+    public static BooleanEncodedValue create() {
+        return new SimpleBooleanEncodedValue(KEY, false);
+    }
+}